// frontend/web/src/services/API/proserve-wb-packaging-api/models/GetRecipeVersionResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { RecipeVersion } from './RecipeVersion';
import {
    RecipeVersionFromJSON,
    RecipeVersionFromJSONTyped,
    RecipeVersionToJSON,
} from './RecipeVersion';

/**
 * 
 * @export
 * @interface GetRecipeVersionResponse
 */
export interface GetRecipeVersionResponse {
    /**
     * 
     * @type {RecipeVersion}
     * @memberof GetRecipeVersionResponse
     */
    recipeVersion: RecipeVersion;
}

/**
 * Check if a given object implements the GetRecipeVersionResponse interface.
 */
export function instanceOfGetRecipeVersionResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "recipeVersion" in value;

    return isInstance;
}

export function GetRecipeVersionResponseFromJSON(json: any): GetRecipeVersionResponse {
    return GetRecipeVersionResponseFromJSONTyped(json, false);
}

export function GetRecipeVersionResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetRecipeVersionResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'recipeVersion': RecipeVersionFromJSON(json['recipe_version']),
    };
}

export function GetRecipeVersionResponseToJSON(value?: GetRecipeVersionResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'recipe_version': RecipeVersionToJSON(value.recipeVersion),
    };
}